SensorSets

Module description

StatFEMEUCLID.SensorSets — Module

SensorSets

This submodule contains structs describing a collection of measurement sensors.

Types

StatFEMEUCLID.SensorSets.AbstractSensorSet — Type

abstract type AbstractSensorSet{T}

StatFEMEUCLID.SensorSets.EqualSensorSet — Type

mutable struct EqualSensorSet{T} <: AbstractSensorSet{T}

A set of sensors with equal measurement accuracies given by a common standard deviation.

Fields

const locations::Vector{Vector{T}} - vector of sensor locations (as coordinate vectors)

const σ::T - common standard deviation

projection::Matrix{T} - Projection matrix from degrees of freedom of an FEM solution to the sensor locations

StatFEMEUCLID.SensorSets.EqualSensorSet — Method

EqualSensorSet(locations::Vector{Vector{T}}, σ::T) where {T<:AbstractFloat}

Construct an EqualSensorSet through given measurement locations and a common standard deviation for all sensors. Note: For use within the statistical FEM, calculate_projection_matrix has to be called as well.

StatFEMEUCLID.SensorSets.FullFieldSensorSet — Type

struct FullFieldSensorSet{T} <: AbstractSensorSet{T}

A mock sensor set when the full FE field without noise is measured. For example when using EUCLID without StatFEM.

Fields

n_dofs::Int{T} - number of unknowns in one dimension

projection::Matrix{T} - Projection matrix from degrees of freedom of an FEM solution to themselves (identity)

StatFEMEUCLID.SensorSets.FullFieldSensorSet — Method

FullFieldSensorSet(n_dofs::Int64)

Construct a FullFieldSensorSet with default matrix value type Float64.

StatFEMEUCLID.SensorSets.FullFieldSensorSet — Method

FullFieldSensorSet(::Type{T}, n_dofs::Int64) where {T <: AbstractFloat}

Construct a FullFieldSensorSet with matrix value type T.

Functions

StatFEMEUCLID.SensorSets.apply_noise! — Method

apply_noise!(sensor_set::EqualSensorSet{T},projected_measurement::Matrix{T};rng::AbstractRNG=default_rng()) where {T}

Sample noise from a Normal distribution with standard deviation sensor_set.σ and add the noise to the measurement at sensor locations.

StatFEMEUCLID.SensorSets.calculate_projection_matrix — Method

calculate_projection_matrix(sensor_set::EqualSensorSet{T}, projection_model::HTTPModel, n_dofs::Int64) where {T}

Query the projection_model to calculate a projection matrix based on the sensor locations of the sensor_set. This matrix maps from the FEM degrees of freedom to the sensor locations

StatFEMEUCLID.SensorSets.project — Method

project(sensor_set::AbstractSensorSet{T},full_field_measurement::Matrix{T}) where {T}

Apply the projection matrix of the sensor_set to the full_field_measurement to obtain a measurement at the sensor locations.

StatFEMEUCLID.SensorSets.projection_matrix — Method

projection_matrix(sensor_set::EqualSensorSet{T}) where T

Returns the projection matrix of the sensor_set mapping from FEM degrees of freedom to the sensor locations.

StatFEMEUCLID.SensorSets.projection_matrix — Method

projection_matrix(sensor_set::FullFieldSensorSet{T}) where {T}

Returns the projection matrix of the sensor_set mapping from FEM degrees of freedom to the sensor locations.