StatFEM

Module description

StatFEMEUCLID.StatFEM — Module

StatFEM

This submodule contains functions for performing the statistical FEM on given measurement data and stochastic prior. The method corrects a given PCE based prior $u_f$ based on given measurements with independent measurement noise using a Gauß-Markov Kalman Filter.

Functions

StatFEMEUCLID.StatFEM.posterior_mean — Method

posterior_mean(sensor_set::AbstractSensorSet{T}, measurements::AbstractMatrix{T}, pce) where {T}

Calculate the posterior mean of a PCE based prior given a sensor_set and measurements.

StatFEMEUCLID.StatFEM.posterior_mean — Method

posterior_mean(sensor_set::AbstractSensorSet{T}, measurements::AbstractVector{T}, pce) where {T}

Calculate the posterior mean of a PCE based prior given a sensor_set and a single measurement.

StatFEMEUCLID.StatFEM.posterior_statistics — Method

posterior_statistics(sensor_set::AbstractSensorSet{T},measurements::AbstractArray{T},pce) where {T}

Calculate the posterior statistics (mean & covariance) of a PCE based prior, given a sensor_set and measurements.

StatFEMEUCLID.StatFEM.posterior_statistics — Method

posterior_statistics(sensor_set::AbstractSensorSet{T},measurements::AbstractVector{T},pce) where {T}

Calculate the posterior statistics (mean & covariance) of a PCE based prior, given a sensor_set and a single measurement vector.