SensorSets

Module description

StatFEMEUCLID.SensorSets — Module

SensorSets

This submodule contains structs describing a collection of measurement sensors. So far, only an EqualSensorSet is implemented, where all sensors have the same measurement accuracy.

Types

StatFEMEUCLID.SensorSets.AbstractSensorSet — Type

abstract type AbstractSensorSet{T}

StatFEMEUCLID.SensorSets.EqualSensorSet — Type

mutable struct EqualSensorSet{T} <: AbstractSensorSet{T}

A set of sensors with equal measurement accuracies given by a common standard deviation.

Fields

const locations::Vector{Vector{T}} - vector of sensor locations (as coordinate vectors)

const σ::T - common standard deviation

projection::Matrix{T} - Projection matrix from degrees of freedom of an FEM solution to the sensor locations

StatFEMEUCLID.SensorSets.EqualSensorSet — Method

EqualSensorSet(locations::Vector{Vector{T}}, σ::T) where {T<:AbstractFloat}

Construct an EqualSensorSet through given measurement locations and a common standard deviation for all sensors. Note: For use within the statistical FEM, calculate_projection_matrix has to be called as well.

Functions

StatFEMEUCLID.SensorSets.calculate_projection_matrix — Method

calculate_projection_matrix(sensor_set::EqualSensorSet{T}, projection_model::HTTPModel, n_dofs::Int64) where {T}

Query the projection_model to calculate a projection matrix based on the sensor locations of the sensor_set. This matrix maps from the FEM degrees of freedom to the sensor locations